fast-forward to present

+++ service/policy/db/namespaces.go
@@ -3,6 +3,7 @@ package db
 import (
 	"context"
 	"database/sql"
+	"fmt"
 	"log/slog"
 	"strings"
 
@@ -11,6 +12,7 @@ import (
 	"github.com/opentdf/platform/protocol/go/common"
 	"github.com/opentdf/platform/protocol/go/policy"
 	"github.com/opentdf/platform/protocol/go/policy/namespaces"
+	"github.com/opentdf/platform/service/internal/logger"
 	"github.com/opentdf/platform/service/pkg/db"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/types/known/wrapperspb"
@@ -21,7 +23,7 @@ type namespaceSelectOptions struct {
 	state   string
 }
 
-func hydrateNamespaceItem(row pgx.Row, opts namespaceSelectOptions) (*policy.Namespace, error) {
+func hydrateNamespaceItem(row pgx.Row, opts namespaceSelectOptions, logger *logger.Logger) (*policy.Namespace, error) {
 	var (
 		id           string
 		name         string
@@ -42,7 +44,7 @@ func hydrateNamespaceItem(row pgx.Row, opts namespaceSelectOptions) (*policy.Nam
 	m := &common.Metadata{}
 	if metadataJSON != nil {
 		if err := protojson.Unmarshal(metadataJSON, m); err != nil {
-			slog.Error("could not unmarshal metadata", slog.String("error", err.Error()))
+			logger.Error("could not unmarshal metadata", slog.String("error", err.Error()))
 			return nil, err
 		}
 	}
@@ -56,11 +58,11 @@ func hydrateNamespaceItem(row pgx.Row, opts namespaceSelectOptions) (*policy.Nam
 	}, nil
 }
 
-func hydrateNamespaceItems(rows pgx.Rows, opts namespaceSelectOptions) ([]*policy.Namespace, error) {
+func hydrateNamespaceItems(rows pgx.Rows, opts namespaceSelectOptions, logger *logger.Logger) ([]*policy.Namespace, error) {
 	var list []*policy.Namespace
 
 	for rows.Next() {
-		n, err := hydrateNamespaceItem(rows, opts)
+		n, err := hydrateNamespaceItem(rows, opts, logger)
 		if err != nil {
 			return nil, err
 		}
@@ -110,12 +112,7 @@ func (c PolicyDBClient) GetNamespace(ctx context.Context, id string) (*policy.Na
 		return nil, err
 	}
 
-	n, err := hydrateNamespaceItem(row, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	return n, nil
+	return hydrateNamespaceItem(row, opts, c.logger)
 }
 
 func listNamespacesSQL(opts namespaceSelectOptions) (string, []interface{}, error) {
@@ -154,20 +151,20 @@ func (c PolicyDBClient) ListNamespaces(ctx context.Context, state string) ([]*po
 
 	sql, args, err := listNamespacesSQL(opts)
 	if err != nil {
-		slog.Error("error listing namespaces", slog.String("error", err.Error()))
+		c.logger.Error("error listing namespaces", slog.String("error", err.Error()))
 		return nil, err
 	}
 
 	rows, err := c.Query(ctx, sql, args)
 	if err != nil {
-		slog.Error("error listing namespaces", slog.String("sql", sql), slog.String("error", err.Error()))
+		c.logger.Error("error listing namespaces", slog.String("sql", sql), slog.String("error", err.Error()))
 		return nil, err
 	}
 	defer rows.Close()
 
-	list, err := hydrateNamespaceItems(rows, opts)
+	list, err := hydrateNamespaceItems(rows, opts, c.logger)
 	if err != nil {
-		slog.Error("error hydrating namespace items", slog.String("sql", sql), slog.String("error", err.Error()))
+		c.logger.Error("error hydrating namespace items", slog.String("sql", sql), slog.String("error", err.Error()))
 		return nil, err
 	}
 
@@ -203,12 +200,13 @@ func (c PolicyDBClient) CreateNamespace(ctx context.Context, r *namespaces.Creat
 		return nil, db.WrapIfKnownInvalidQueryErr(e)
 	}
 
-	if err = unmarshalMetadata(metadataJSON, m); err != nil {
+	if err = unmarshalMetadata(metadataJSON, m, c.logger); err != nil {
 		return nil, err
 	}
 
 	// Update FQN
-	c.upsertAttrFqn(ctx, attrFqnUpsertOptions{namespaceID: id})
+	fqn := c.upsertAttrFqn(ctx, attrFqnUpsertOptions{namespaceID: id})
+	c.logger.Debug("upserted fqn for created namespace", slog.Any("fqn", fqn))
 
 	return &policy.Namespace{
 		Id:       id,
@@ -275,7 +273,7 @@ func unsafeUpdateNamespaceSQL(id string, name string) (string, []interface{}, er
 }
 
 func (c PolicyDBClient) UnsafeUpdateNamespace(ctx context.Context, id string, name string) (*policy.Namespace, error) {
-	sql, args, err := unsafeUpdateNamespaceSQL(id, name)
+	sql, args, err := unsafeUpdateNamespaceSQL(id, strings.ToLower(name))
 	if db.IsQueryBuilderSetClauseError(err) {
 		return &policy.Namespace{
 			Id: id,
@@ -289,10 +287,25 @@ func (c PolicyDBClient) UnsafeUpdateNamespace(ctx context.Context, id string, na
 	if err != nil {
 		return nil, err
 	}
-	// Update FQN
-	c.upsertAttrFqn(ctx, attrFqnUpsertOptions{namespaceID: id})
 
-	return hydrateNamespaceItem(row, namespaceSelectOptions{})
+	// Update all FQNs that may contain the namespace name
+	fqn := c.upsertAttrFqn(ctx, attrFqnUpsertOptions{namespaceID: id})
+	c.logger.Debug("upserted fqn for unsafely updated namespace", slog.Any("fqn", fqn))
+
+	attrs, err := c.ListAllAttributes(ctx, StateAny, id)
+	if err != nil {
+		return nil, err
+	}
+	for _, attr := range attrs {
+		fqn = c.upsertAttrFqn(ctx, attrFqnUpsertOptions{namespaceID: id, attributeID: attr.GetId()})
+		c.logger.Debug("upserted definition fqn for unsafely updated namespace", slog.Any("fqn", fqn))
+		for _, value := range attr.GetValues() {
+			fqn = c.upsertAttrFqn(ctx, attrFqnUpsertOptions{namespaceID: id, attributeID: attr.GetId(), valueID: value.GetId()})
+			c.logger.Debug("upserted value fqn for unsafely updated namespace", slog.Any("fqn", fqn))
+		}
+	}
+
+	return hydrateNamespaceItem(row, namespaceSelectOptions{}, c.logger)
 }
 
 func setNamespaceActiveStateSQL(id string, isActive bool) (string, []interface{}, error) {
@@ -320,7 +333,7 @@ func (c PolicyDBClient) DeactivateNamespace(ctx context.Context, id string) (*po
 	}
 
 	if !allAttrsDeactivated {
-		slog.Warn("deactivating the namespace with existed attributes can affect access to related data. Please be aware and proceed accordingly.")
+		c.logger.Warn("deactivating the namespace with existed attributes can affect access to related data. Please be aware and proceed accordingly.")
 	}
 
 	sql, args, err := setNamespaceActiveStateSQL(id, false)
@@ -341,7 +354,7 @@ func (c PolicyDBClient) UnsafeReactivateNamespace(ctx context.Context, id string
 	}
 
 	if len(attrs) > 0 {
-		slog.Warn("reactivating the namespace with existing attributes can affect access to related data. Please be aware and proceed accordingly.")
+		c.logger.Warn("reactivating the namespace with existing attributes can affect access to related data. Please be aware and proceed accordingly.")
 	}
 
 	sql, args, err := setNamespaceActiveStateSQL(id, true)
@@ -364,7 +377,15 @@ func deleteNamespaceSQL(id string) (string, []interface{}, error) {
 		ToSql()
 }
 
-func (c PolicyDBClient) UnsafeDeleteNamespace(ctx context.Context, id string) (*policy.Namespace, error) {
+func (c PolicyDBClient) UnsafeDeleteNamespace(ctx context.Context, existing *policy.Namespace, fqn string) (*policy.Namespace, error) {
+	if existing == nil {
+		return nil, fmt.Errorf("namespace not found: %w", db.ErrNotFound)
+	}
+	id := existing.GetId()
+
+	if existing.GetFqn() != fqn {
+		return nil, fmt.Errorf("fqn mismatch: %w", db.ErrNotFound)
+	}
 	sql, args, err := deleteNamespaceSQL(id)
 	if err != nil {
 		return nil, err
